% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phangornMCMC.R
\name{phangornMCMC}
\alias{phangornMCMC}
\title{phangornMCMC
This function runs a Markov chain Monte Carlo (MCMC) algorithm to generate a
set of trees which is returned with their likelihoods.}
\usage{
phangornMCMC(x, ntrees = 3000, burnin = 1000, frequency = 1,
  tree0 = NULL, model = NULL, printevery = 100, bf = baseFreq(x),
  Q = rep(1, 6), optBf = TRUE, optQ = TRUE, rooted = TRUE)
}
\arguments{
\item{x}{a set of DNA sequences, typically an object of class "DNAbin" or "phyDat".}

\item{ntrees}{the number of trees to output.}

\item{burnin}{the number of trees to discard as “burn-in”.}

\item{frequency}{the frequency at which trees are sampled.}

\item{tree0}{the initial tree of the chain; by default, a fastME / UPGMA (un-) tree
from a JC69 distance is generated.}

\item{model}{the transition model}

\item{printevery}{an integer specifying the frequency at which to print the numbers of trees proposed and accepted; set to 0 to cancel all printings.}

\item{bf}{base frequencies}

\item{Q}{rate matrix}
}
\description{
phangornMCMC
This function runs a Markov chain Monte Carlo (MCMC) algorithm to generate a
set of trees which is returned with their likelihoods.
}
